// src/app/users/models/user.ts
export interface User {
  id: string;
  email: string;
  firstname?: string;
  lastname?: string;
  birthdate?: string | null;
  phone?: string;
  nationality?: string;
  age?: number | string;
  isActive?: boolean;
  role_id?: string;
  role?: {
    id: string;
    rolename: string;
  };
  courses?: any[];
  createdAt?: Date | string;
  updatedAt?: Date | string;
}